💄 색상리스트 컴포넌트로 분리
